+++ Dissects/geometry/__init__.py
@@ -1,1 +1,1 @@
-from .skel import *
+from .skeleton import Skeleton